2/3 improvements, 1/3 regressions

+++ util/src/offsetmeasurement.cpp
@@ -4,10 +4,14 @@
 #include <string>
 
 OffsetMeasurement::OffsetMeasurement(int index, int64_t starttime_ns, int64_t endtime_ns, int samples_sent,
-                                     int samples, int used, int64_t avg_ns, bool valid)
-    : m_index(index), m_starttime_ns(starttime_ns), m_endtime_ns(endtime_ns),
-      m_sentSamples(samples_sent), m_collectedSamples(samples), m_usedSamples(used),
-      m_offset_ns(avg_ns), m_valid(valid)
+                                     int samples, size_t used, int64_t avg_ns, bool valid)
+    : m_index(index),
+      m_starttime_ns(starttime_ns), m_endtime_ns(endtime_ns),
+      m_sentSamples(samples_sent),
+      m_collectedSamples(samples),
+      m_usedSamples(used),
+      m_offset_ns(avg_ns),
+      m_valid(valid)
 {
 }
 
@@ -32,6 +36,7 @@ std::string OffsetMeasurement::toString() const
                        m_offset_ns/1000.0, m_ppm,
                        m_valid ? "" : "invalid");
 }
+
 
 double OffsetMeasurement::packageLossPct() const
 {
